refactor: moving zones into the common schema/azure directory
